+++ dgisim/src/character/character.py
@@ -35,6 +35,7 @@ __all__ = [
     "Bennett",
     "ElectroHypostasis",
     "FatuiPyroAgent",
+    "Jean",
     "KaedeharaKazuha",
     "Kaeya",
     "Keqing",
@@ -793,6 +794,85 @@ class FatuiPyroAgent(Character):
             energy=0,
             max_energy=2,
             hiddens=stts.Statuses((stt.StealthMasterStatus(),)),
+            equipments=stts.EquipmentStatuses(()),
+            statuses=stts.Statuses(()),
+            elemental_aura=ElementalAura.from_default(),
+        )
+
+
+class Jean(Character):
+    _ELEMENT = Element.ANEMO
+    _WEAPON_TYPE = WeaponType.SWORD
+    _TALENT_STATUS = stt.LandsOfDandelionStatus
+    _FACTIONS = frozenset((Faction.MONDSTADT,))
+
+    _NORMAL_ATTACK_COST = AbstractDices({
+        Element.ANEMO: 1,
+        Element.ANY: 2,
+    })
+    _ELEMENTAL_SKILL1_COST = AbstractDices({
+        Element.ANEMO: 3,
+    })
+    _ELEMENTAL_BURST_COST = AbstractDices({
+        Element.ANEMO: 4,
+    })
+
+    @override
+    def _normal_attack(self, game_state: GameState, source: StaticTarget) -> tuple[eft.Effect, ...]:
+        return normal_attack_template(
+            game_state=game_state,
+            source=source,
+            element=Element.PHYSICAL,
+            damage=2,
+        )
+
+    @override
+    def _elemental_skill1(self, game_state: GameState, source: StaticTarget) -> tuple[eft.Effect, ...]:
+        return (
+            eft.ReferredDamageEffect(
+                source=source,
+                target=DynamicCharacterTarget.OPPO_ACTIVE,
+                element=Element.ANEMO,
+                damage=3,
+                damage_type=DamageType(elemental_skill=True),
+            ),
+            eft.ForwardSwapCharacterEffect(
+                target_player=source.pid.other(),
+            ),
+        )
+
+    @override
+    def _elemental_burst(self, game_state: GameState, source: StaticTarget) -> tuple[eft.Effect, ...]:
+        effects: list[eft.Effect] = [
+            eft.EnergyDrainEffect(
+                target=source,
+                drain=self.get_max_energy(),
+            ),
+        ]
+        self_characters = game_state.get_player(source.pid).get_characters()
+        for character in self_characters.get_alive_character_in_activity_order():
+            effects.append(eft.RecoverHPEffect(
+                target=StaticTarget.from_char_id(source.pid, character.get_id()),
+                recovery=2,
+            ))
+        effects.append(
+            eft.AddSummonEffect(
+                target_pid=source.pid,
+                summon=sm.DandelionFieldSummon,
+            ),
+        )
+        return tuple(effects)
+
+    @classmethod
+    def from_default(cls, id: int = -1) -> Self:
+        return cls(
+            id=id,
+            alive=True,
+            hp=10,
+            max_hp=10,
+            energy=0,
+            max_energy=3,
+            hiddens=stts.Statuses(()),
             equipments=stts.EquipmentStatuses(()),
             statuses=stts.Statuses(()),
             elemental_aura=ElementalAura.from_default(),